move some multiprocessing boilerplate into base sim class
